Move list binding functions into helper class
